@@ -1,8 +1,9 @@
 {
   "name": "js-e2e-express-server",
-  "version": "1.0.0",
-  "description": "JavaScript server written with Express.js",
+  "version": "2.0.0",
+  "description": "ESM version of JavaScript server written with Express.js",
   "main": "index.js",
+  "type": "module",
   "directories": {
     "doc": "docs",
     "test": "test"
@@ -11,7 +12,7 @@
     "build": "npm run lint",
     "start": "node src/index.js",
     "start:dev": "npm run lint && nodemon --exec node src/index.js --unhandled-rejections=warn",
-    "test": "jest --detectOpenHandles --coverage",
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js --detectOpenHandles --coverage",
     "lint": "eslint src/**/*.js --fix"
   },
   "repository": {
@@ -26,20 +27,20 @@
   },
   "homepage": "https://github.com/azure-samples/js-e2e-express-server#readme",
   "devDependencies": {
-    "eslint": "^7.9.0",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-airbnb-base": "^14.2.0",
-    "eslint-config-prettier": "^6.11.0",
+    "eslint": "^8.56.0",
+    "eslint-config-airbnb": "^19.0.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "jest": "^26.4.2",
-    "nodemon": "^2.0.4",
-    "prettier": "^2.1.2",
+    "eslint-plugin-prettier": "^5.1.3",
+    "jest": "^29.7.0",
+    "nodemon": "^3.0.3",
+    "prettier": "^3.2.4",
     "supertest": "^6.1.1"
   },
   "dependencies": {
     "body-parser": "^1.19.0",
-    "dotenv": "^8.2.0",
+    "dotenv": "^16.3.1",
     "express": "^4.17.1",
     "serve-favicon": "^2.5.0"
   },
